refactor default params mixin, add support for aiomysql and aiosqlite
